+++ src/components/image-image/image-image.interface.ts
@@ -1,6 +1,0 @@
-export interface ImageImageProps {
-  smallImageSrc?: string;
-  smallImageAlt?: string;
-  largeImageSrc: string;
-  largeImageAlt?: string;
-}