+++ src/components/media-text-component/index.ts
@@ -1,0 +1,2 @@
+export { MediaTextComponent } from './MediaTextComponent';
+export type { MediaTextComponentProps } from './media-text-component.interface';