+++ src/components/media-text-component/media-text-component.interface.ts
@@ -1,0 +1,14 @@
+import { LinkProps } from "@zepdev/design-system-component-library-react";
+
+export interface MediaTextComponentProps {
+    mediaUrl: string;
+    mediaType: 'image' | 'video';
+    imageOrientation?: 'horizontal' | 'vertical';
+    imageAlt?: string;
+    mediaPosition: 'left' | 'right';
+    headline: string;
+    description?: string;
+    buttonText?: string;
+    links?: LinkProps[];
+    videoThumbnail?: string;
+  }