+++ src/components/media-text-small/index.ts
@@ -1,0 +1,2 @@
+export {MediaTextSmall} from "./mediaTextSmall";
+export type {MediaTextSmallProps} from "./media-text-small.interface";