+++ src/components/product-highlight/product-highlight.interface.ts
@@ -1,9 +1,9 @@
 import { BlocksContent } from '@strapi/blocks-react-renderer';
 import { MouseEvent } from 'react';
+import { FunctionalIconNames } from '~/@zepdev/design-system-component-library-react';
 import { GlobalVariants } from '../../interfaces/global-variants';
 import { HeaderLongProps } from '../header-long';
-import { ZpsButtonVariant } from '../zps-button';
-import { FunctionalIconNames } from '~/@zepdev/design-system-component-library-react';
+import { ZsdButtonVariant } from '../zsd-button';
 
 export interface ProductHighLightProps extends Partial<Omit<HeaderLongProps, 'variant'>> {
   imageAlignment?: 'left' | 'right';
@@ -11,12 +11,12 @@ export interface ProductHighLightProps extends Partial<Omit<HeaderLongProps, 'va
   imageAlt: string;
   productDetails: BlocksContent[];
   buttonText?: string;
-  buttonVariant?: ZpsButtonVariant | 'link';
+  buttonVariant?: ZsdButtonVariant | 'link';
   onClick?: (ev: MouseEvent<HTMLElement>) => void;
   variant?: GlobalVariants;
   id?: string;
   buttonUrl?: string;
-  buttonType?: 'download' | 'default'
-  buttonIcon?: FunctionalIconNames,
-  buttonIconPosition?: 'left' | 'right',
+  buttonType?: 'download' | 'default';
+  buttonIcon?: FunctionalIconNames;
+  buttonIconPosition?: 'left' | 'right';
 }