+++ src/components/used-machine-info/UsedMachineInfo.interface.ts
@@ -1,9 +1,0 @@
-import { MouseEvent } from 'react';
-import { GalleryProps } from '../gallery';
-import { UsedMachineDescriptionProps } from '../used-machine-pattern';
-
-export interface UsedMachineInfoProps extends GalleryProps, UsedMachineDescriptionProps {
-  onExposeClick: (ev: MouseEvent<HTMLElement>) => void;
-  onOfferClick: (ev: MouseEvent<HTMLElement>) => void;
-  description: string,
-}