+++ src/components/used-machine-info/index.ts
@@ -1,2 +1,0 @@
-export * from './UsedMachineInfo';
-export { type UsedMachineInfoProps } from './UsedMachineInfo.interface';