+++ src/components/zsd-button/index.ts
@@ -1,3 +1,3 @@
-export { Button } from './ZsdButton';
 export { ZsdButtonVariant } from './zsd-button.interface';
-export type { ZpsButtonProps } from './zsd-button.interface';
+export type { ZsdButtonProps } from './zsd-button.interface';
+export { Button } from './ZsdButton';